+++ frontend/components/common/badge/index.tsx
@@ -1,10 +1,10 @@
-import Text from '../text';
+import { Text } from '@/components/common';
 
 interface Props {
   content: string;
 }
 
-export default function Badge({ content }: Props) {
+export function Badge({ content }: Props) {
   return (
     <div className='px-3 py-1 rounded-xl bg-default-color bg-opacity-10'>
       <Text variant='semi-base' content={content} />
