+++ frontend/components/common/button/index.tsx
@@ -1,5 +1,5 @@
 import { ComponentPropsWithRef } from 'react';
 
-export default function Button({ children, ...rest }: ComponentPropsWithRef<'button'>) {
+export function Button({ children, ...rest }: ComponentPropsWithRef<'button'>) {
   return <button {...rest}>{children}</button>;
 }
