+++ frontend/components/common/index.ts
@@ -1,0 +1,11 @@
+export { Badge } from './badge';
+export { Button } from './button';
+export { Checkbox } from './checkbox';
+export { Input } from './input';
+export { Layout } from './layout';
+export { Modal } from './modal';
+export { Radio } from './radio';
+export { Select } from './select';
+export { Slider } from './slider';
+export { Spinner } from './spinner';
+export { Text } from './text';
