+++ frontend/components/common/input/index.tsx
@@ -1,8 +1,7 @@
 import Image from 'next/image';
 import { ComponentPropsWithoutRef, KeyboardEventHandler, MouseEventHandler } from 'react';
 
-import Button from '../button';
-import Text from '../text';
+import { Button, Text } from '@/components/common';
 
 interface Props extends ComponentPropsWithoutRef<'input'> {
   label?: string;
@@ -12,7 +11,7 @@ interface Props extends ComponentPropsWithoutRef<'input'> {
   onClickButton?: () => void;
 }
 
-export default function Input({
+export function Input({
   label = '',
   value,
   hasButton = false,
