+++ frontend/components/common/layout/index.tsx
@@ -1,10 +1,10 @@
 import { PropsWithChildren } from 'react';
 
-import Footer from './footer';
-import Header from './header';
-import Main from './main';
+import { Footer } from './footer';
+import { Header } from './header';
+import { Main } from './main';
 
-export default function Layout({ children }: PropsWithChildren) {
+export function Layout({ children }: PropsWithChildren) {
   return (
     <div className='flex flex-col flex-1'>
       <Header />
