+++ frontend/components/common/layout/main.tsx
@@ -1,5 +1,5 @@
 import { PropsWithChildren } from 'react';
 
-export default function Main({ children }: PropsWithChildren) {
+export function Main({ children }: PropsWithChildren) {
   return <main className='flex flex-1'>{children}</main>;
 }
