+++ frontend/components/common/select/context/index.tsx
@@ -6,6 +6,6 @@ interface Props extends PropsWithChildren {
   value: SelectContextValue;
 }
 
-export default function SelectContextProvider({ value, children }: Props) {
+export function SelectContextProvider({ value, children }: Props) {
   return <SelectContext.Provider value={value}>{children}</SelectContext.Provider>;
 }
