+++ frontend/components/common/spinner/index.tsx
@@ -1,3 +1,3 @@
-export default function Spinner() {
+export function Spinner() {
   return <div className='w-12 h-12 border-8 border-t-green-800 rounded-full animate-spin' />;
 }
