+++ frontend/components/filter/jobsFilter/components/applyJobsButton.tsx
@@ -2,8 +2,7 @@ import { Job } from '@/types/api/jobs';
 
 import { PARAMS } from '@/constants/api/queryParams';
 
-import Button from '@/components/common/button';
-import Text from '@/components/common/text';
+import { Button, Text } from '@/components/common';
 
 import { usePushRouteWithQueryParam } from '@/hooks/usePushRouteWithQueryParam';
