+++ frontend/components/filter/jobsFilter/components/resetJobsButton.tsx
@@ -1,5 +1,4 @@
-import Button from '@/components/common/button';
-import Text from '@/components/common/text';
+import { Button, Text } from '@/components/common';
 
 interface Props {
   clearSelectedCategory: () => void;
