+++ frontend/components/filter/locationFilter/citiesCheckbox/index.tsx
@@ -4,7 +4,7 @@ import { District } from '@/types/api/location';
 
 import { LOCATIONS } from '@/constants/components/location';
 
-import Checkbox from '@/components/common/checkbox';
+import { Checkbox } from '@/components/common';
 
 interface Props {
   cities: string[] | null;
