+++ frontend/components/filter/serviceTypesFilter/index.tsx
@@ -1,7 +1,6 @@
 import { PARAMS, SERVICE_TYPES } from '@/constants/api/queryParams';
 
-import Radio from '@/components/common/radio';
-import Text from '@/components/common/text';
+import { Radio, Text } from '@/components/common';
 
 import { usePushRouteWithQueryParam } from '@/hooks/usePushRouteWithQueryParam';
