+++ frontend/types/components/common.ts
@@ -1,0 +1,1 @@
+export type Variant = 'title' | 'middle-title' | 'semi-title' | 'full-base' | 'base' | 'semi-base';
