+++ frontend/types/components/index.ts
@@ -1,0 +1,1 @@
+export type { Variant } from './common';
